Django AUTH

+++ Web/04_django/Corona/CORONA03/articles/views.py
@@ -1,5 +1,7 @@
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render, redirect, get_object_or_404, HttpResponse
+# from django.http import HttpResponse
 from django.views.decorators.http import require_http_methods, require_POST
+from django.contrib.auth.decorators import login_required
 from .models import Article
 from .forms import ArticleForm
 
@@ -11,6 +13,7 @@ def index(request):
     return render(request, 'articles/index.html', context)
 
 
+@login_required
 @require_http_methods(['GET', 'POST'])
 def new(request):
     if request.method == 'POST':
@@ -41,6 +44,7 @@ def detail(request, pk):
     return render(request, 'articles/detail.html', context)
 
 
+@login_required
 @require_http_methods(['GET', 'POST'])
 def edit(request, pk):
     article = get_object_or_404(Article, pk=pk)
@@ -66,8 +70,11 @@ def edit(request, pk):
     return render(request, 'articles/form.html', context)
 
 
+# @login_required
 @require_POST
 def delete(request, pk):
-    article = get_object_or_404(Article, pk=pk)
-    article.delete()
-    return redirect('articles:index')
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=pk)
+        article.delete()
+        return redirect('articles:index')
+    return HttpResponse('You are Unauthorized', status=401)
